Restructures `preprocess` to include building block, extensible transforms (#4659)
The current `devices.qubit.preprocess` module has several transforms
optimized for default qubit. Other devices would need to rewrite similar
code in order to extend the behavior.

This PR restructures the module so that `devices.preprocess` now offers
the following transforms:

* `decompose`: this accepts a stopping condition and decomposes
operations until the stopping condition is met
* `validate_measurements`: this accepts a stopping_condition for
observables. It validates all observables are accepted, only sample
based measurements are present when sampling, and only state based
measurements are present when analytic
* `validate_device_wires`: This makes sure device wire constraints are
met, and fills in all wires for measurements like `StateMP` and
`ProbabilityMP` that act on all available wires.
* `validate_multiprocessing_workers`: Checks that the CPU is not
oversubscribed for a given number of workers.
* `warn_about_trainable_observables`: raises a warning is any observable
are trainable


For example, a plugin with a list of supported operations can use the
`decompose` transform by doing:

```
def stopping_condition(op: qml.operation.Operator) -> bool:
    return op.name in supported_operations

transform_program.add_transform(decompose, stopping_condition=stopping_condition)

```

**Overview of all the changes:**

So this PR ended up changing a lot of lines of code, but most of them
are moving tests around.


As the tests for default qubit started being quite extensive, I added a
folder `tests/devices/default_qubit` and moved the default qubit tests
there over two files.

Any test for default qubit specific preprocessing went into
`tests/devices/default_qubit/test_default_qubit_preprocessing.py`.

More general preprocessing tests are in
`tests/devices/test_preprocessing.py`.

Since the transforms are no longer default qubit specific, they are
moved out of the `qubit` folder.


Since `DefaultQubit.preprocess` was getting rather long, I split out two
private helper methods, `_add_adjoint_transforms` and
`_setup_execution_config`. This change just keeps
`DefaultQubit.preprocess` a little more manage able.

---------
