+++ pennylane/devices/qubit/__init__.py
@@ -26,7 +26,6 @@ at your own discretion.
     measure
     measure_with_samples
     sample_state
-    preprocess
     simulate
     adjoint_jacobian
     adjoint_jvp
@@ -37,6 +36,5 @@ from .apply_operation import apply_operation
 from .adjoint_jacobian import adjoint_jacobian, adjoint_jvp, adjoint_vjp
 from .initialize_state import create_initial_state
 from .measure import measure
-from .preprocess import preprocess
 from .sampling import sample_state, measure_with_samples
 from .simulate import simulate, get_final_state, measure_final_state
