+++ pennylane/transforms/defer_measurements.py
@@ -26,6 +26,13 @@ from pennylane.queuing import QueuingManager
 # pylint: disable=too-many-branches
 
 
+def null_postprocessing(results):
+    """A postprocessing function returned by a transform that only converts the batch of results
+    into a result for a single ``QuantumTape``.
+    """
+    return results[0]
+
+
 @transform
 def defer_measurements(tape: QuantumTape) -> (Sequence[QuantumTape], Callable):
     """Quantum function transform that substitutes operations conditioned on
@@ -130,10 +137,16 @@ def defer_measurements(tape: QuantumTape) -> (Sequence[QuantumTape], Callable):
     tensor([0.76960924, 0.13204407, 0.08394415, 0.01440254], requires_grad=True)
     """
     # pylint: disable=protected-access
+    if not any(isinstance(o, MidMeasureMP) for o in tape.operations):
+        return (tape,), null_postprocessing
 
     cv_types = (qml.operation.CVOperation, qml.operation.CVObservable)
-    ops_cv = any(isinstance(op, cv_types) for op in tape.operations)
-    obs_cv = any(isinstance(getattr(op, "obs", None), cv_types) for op in tape.measurements)
+    ops_cv = any(isinstance(op, cv_types) and op.name != "Identity" for op in tape.operations)
+    obs_cv = any(
+        isinstance(getattr(op, "obs", None), cv_types)
+        and not isinstance(getattr(op, "obs", None), qml.Identity)
+        for op in tape.measurements
+    )
     if ops_cv or obs_cv:
         raise ValueError("Continuous variable operations and observables are not supported.")
 
@@ -199,12 +212,6 @@ def defer_measurements(tape: QuantumTape) -> (Sequence[QuantumTape], Callable):
         new_measurements.append(mp)
 
     new_tape = type(tape)(new_operations, new_measurements, shots=tape.shots)
-
-    def null_postprocessing(results):
-        """A postprocesing function returned by a transform that only converts the batch of results
-        into a result for a single ``QuantumTape``.
-        """
-        return results[0]
 
     return [new_tape], null_postprocessing
