+++ tests/devices/default_qubit/test_default_qubit.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for default qubit 2."""
+"""Tests for default qubit."""
 # pylint: disable=import-outside-toplevel, no-member
 
 import pytest
@@ -19,10 +19,8 @@ import pytest
 import numpy as np
 
 import pennylane as qml
-from pennylane.measurements import SampleMP, StateMP, ProbabilityMP
 
 from pennylane.devices import DefaultQubit, ExecutionConfig
-from pennylane.devices.qubit.preprocess import validate_and_expand_adjoint
 
 
 def test_name():
@@ -58,135 +56,23 @@ def test_debugger_attribute():
     assert dev._debugger is None
 
 
-class TestSnapshotMulti:
-    def test_snapshot_multiprocessing_execute(self):
-        """DefaultQubit cannot execute tapes with Snapshot if `max_workers` is not `None`"""
-        dev = DefaultQubit(max_workers=2)
+def test_snapshot_multiprocessing_qnode():
+    """DefaultQubit cannot execute tapes with Snapshot if `max_workers` is not `None`"""
+    dev = DefaultQubit(max_workers=2)
 
-        tape = qml.tape.QuantumScript(
-            [
-                qml.Snapshot(),
-                qml.Hadamard(wires=0),
-                qml.Snapshot("very_important_state"),
-                qml.CNOT(wires=[0, 1]),
-                qml.Snapshot(),
-            ],
-            [qml.expval(qml.PauliX(0))],
-        )
-        with pytest.raises(
-            RuntimeError, match="ProcessPoolExecutor cannot execute a QuantumScript"
-        ):
-            program, _ = dev.preprocess()
-            program([tape])
+    @qml.qnode(dev)
+    def circuit():
+        qml.Snapshot("tag")
+        qml.Hadamard(wires=0)
+        qml.CNOT(wires=[0, 1])
+        qml.Snapshot()
+        return qml.expval(qml.PauliX(0) + qml.PauliY(0))
 
-    def test_snapshot_multiprocessing_qnode(self):
-        """DefaultQubit cannot execute tapes with Snapshot if `max_workers` is not `None`"""
-        dev = DefaultQubit(max_workers=2)
-
-        @qml.qnode(dev)
-        def circuit():
-            qml.Snapshot("tag")
-            qml.Hadamard(wires=0)
-            qml.CNOT(wires=[0, 1])
-            qml.Snapshot()
-            return qml.expval(qml.PauliX(0) + qml.PauliY(0))
-
-        with pytest.raises(
-            qml.DeviceError,
-            match="Debugging with ``Snapshots`` is not available with multiprocessing.",
-        ):
-            qml.snapshots(circuit)()
-
-
-# pylint: disable=too-few-public-methods
-class TestPreprocessing:
-    """Unit tests for the preprocessing method."""
-
-    def test_chooses_best_gradient_method(self):
-        """Test that preprocessing chooses backprop as the best gradient method."""
-        dev = DefaultQubit()
-
-        config = ExecutionConfig(
-            gradient_method="best", use_device_gradient=None, grad_on_execution=None
-        )
-
-        _, new_config = dev.preprocess(config)
-
-        assert new_config.gradient_method == "backprop"
-        assert new_config.use_device_gradient
-        assert not new_config.grad_on_execution
-
-    def test_config_choices_for_adjoint(self):
-        """Test that preprocessing request grad on execution and says to use the device gradient if adjoint is requested."""
-        dev = DefaultQubit()
-
-        config = ExecutionConfig(
-            gradient_method="adjoint", use_device_gradient=None, grad_on_execution=None
-        )
-
-        _, new_config = dev.preprocess(config)
-
-        assert new_config.use_device_gradient
-        assert new_config.grad_on_execution
-
-    @pytest.mark.parametrize("max_workers", [None, 1, 2, 3])
-    def test_config_choices_for_threading(self, max_workers):
-        """Test that preprocessing request grad on execution and says to use the device gradient if adjoint is requested."""
-        dev = DefaultQubit()
-
-        config = ExecutionConfig(device_options={"max_workers": max_workers})
-        _, new_config = dev.preprocess(config)
-
-        assert new_config.device_options["max_workers"] == max_workers
-
-    def test_circuit_wire_validation(self):
-        """Test that preprocessing validates wires on the circuits being executed."""
-        dev = DefaultQubit(wires=3)
-        circuit_valid_0 = qml.tape.QuantumScript([qml.PauliX(0)])
-        program, _ = dev.preprocess()
-        circuits, _ = program([circuit_valid_0])
-        assert circuits == (circuit_valid_0,)
-
-        circuit_valid_1 = qml.tape.QuantumScript([qml.PauliX(1)])
-        program, _ = dev.preprocess()
-        circuits, _ = program([circuit_valid_0, circuit_valid_1])
-        assert circuits == tuple([circuit_valid_0, circuit_valid_1])
-
-        invalid_circuit = qml.tape.QuantumScript([qml.PauliX(4)])
-        with pytest.raises(qml.wires.WireError, match=r"Cannot run circuit\(s\) on"):
-            program, _ = dev.preprocess()
-            program(
-                [
-                    invalid_circuit,
-                ]
-            )
-
-        with pytest.raises(qml.wires.WireError, match=r"Cannot run circuit\(s\) on"):
-            program, _ = dev.preprocess()
-            program([circuit_valid_0, invalid_circuit])
-
-    @pytest.mark.parametrize(
-        "mp_fn,mp_cls,shots",
-        [
-            (qml.sample, SampleMP, 10),
-            (qml.state, StateMP, None),
-            (qml.probs, ProbabilityMP, None),
-        ],
-    )
-    def test_measurement_is_swapped_out(self, mp_fn, mp_cls, shots):
-        """Test that preprocessing swaps out any MP with no wires or obs"""
-        dev = DefaultQubit(wires=3)
-        original_mp = mp_fn()
-        exp_z = qml.expval(qml.PauliZ(0))
-        qs = qml.tape.QuantumScript([qml.Hadamard(0)], [original_mp, exp_z], shots=shots)
-        program, _ = dev.preprocess()
-        tapes, _ = program([qs])
-        assert len(tapes) == 1
-        tape = tapes[0]
-        assert tape.operations == qs.operations
-        assert tape.measurements != qs.measurements
-        assert qml.equal(tape.measurements[0], mp_cls(wires=[0, 1, 2]))
-        assert tape.measurements[1] is exp_z
+    with pytest.raises(
+        qml.DeviceError,
+        match="Debugging with ``Snapshots`` is not available with multiprocessing.",
+    ):
+        qml.snapshots(circuit)()
 
 
 class TestSupportsDerivatives:
@@ -1009,8 +895,10 @@ class TestAdjointDifferentiation:
         dev = DefaultQubit(max_workers=max_workers)
         x = np.array(np.pi / 7)
         qs = qml.tape.QuantumScript([qml.RX(x, 0)], [qml.expval(qml.PauliZ(0))])
-        qs, _ = validate_and_expand_adjoint(qs)
-        qs = qs[0]
+
+        config = ExecutionConfig(gradient_method="adjoint")
+        batch, _ = dev.preprocess(config)[0]((qs,))
+        qs = batch[0]
         expected_grad = -qml.math.sin(x)
         actual_grad = dev.compute_derivatives(qs, self.ec)
         assert isinstance(actual_grad, np.ndarray)
@@ -1027,8 +915,9 @@ class TestAdjointDifferentiation:
         dev = DefaultQubit(max_workers=max_workers)
         x = np.array(np.pi / 7)
         qs = qml.tape.QuantumScript([qml.RX(x, 0)], [qml.expval(qml.PauliZ(0))])
-        qs, _ = validate_and_expand_adjoint(qs)
-        qs = qs[0]
+        config = ExecutionConfig(gradient_method="adjoint")
+        batch, _ = dev.preprocess(config)[0]((qs,))
+        qs = batch[0]
         expected_grad = -qml.math.sin(x)
         actual_grad = dev.compute_derivatives([qs], self.ec)
         assert isinstance(actual_grad, tuple)
@@ -1083,8 +972,9 @@ class TestAdjointDifferentiation:
 
         qs = qml.tape.QuantumScript([qml.RX(x, 0)], [qml.expval(qml.PauliZ(0))])
 
-        qs, _ = validate_and_expand_adjoint(qs)
-        qs = qs[0]
+        config = ExecutionConfig(gradient_method="adjoint")
+        batch, _ = dev.preprocess(config)[0]((qs,))
+        qs = batch[0]
 
         expected_grad = -qml.math.sin(x) * tangent[0]
         actual_grad = dev.compute_jvp(qs, tangent, self.ec)
@@ -1105,8 +995,9 @@ class TestAdjointDifferentiation:
 
         qs = qml.tape.QuantumScript([qml.RX(x, 0)], [qml.expval(qml.PauliZ(0))])
 
-        qs, _ = validate_and_expand_adjoint(qs)
-        qs = qs[0]
+        config = ExecutionConfig(gradient_method="adjoint")
+        batch, _ = dev.preprocess(config)[0]((qs,))
+        qs = batch[0]
 
         expected_grad = -qml.math.sin(x) * tangent[0]
         actual_grad = dev.compute_jvp([qs], [tangent], self.ec)
@@ -1180,8 +1071,9 @@ class TestAdjointDifferentiation:
         cotangent = (0.456,)
 
         qs = qml.tape.QuantumScript([qml.RX(x, 0)], [qml.expval(qml.PauliZ(0))])
-        qs, _ = validate_and_expand_adjoint(qs)
-        qs = qs[0]
+        config = ExecutionConfig(gradient_method="adjoint")
+        batch, _ = dev.preprocess(config)[0]((qs,))
+        qs = batch[0]
 
         expected_grad = -qml.math.sin(x) * cotangent[0]
         actual_grad = dev.compute_vjp(qs, cotangent, self.ec)
@@ -1201,8 +1093,9 @@ class TestAdjointDifferentiation:
         cotangent = (0.456,)
 
         qs = qml.tape.QuantumScript([qml.RX(x, 0)], [qml.expval(qml.PauliZ(0))])
-        qs, _ = validate_and_expand_adjoint(qs)
-        qs = qs[0]
+        config = ExecutionConfig(gradient_method="adjoint")
+        batch, _ = dev.preprocess(config)[0]((qs,))
+        qs = batch[0]
 
         expected_grad = -qml.math.sin(x) * cotangent[0]
         actual_grad = dev.compute_vjp([qs], [cotangent], self.ec)
@@ -1267,131 +1160,6 @@ class TestAdjointDifferentiation:
 
         assert np.isclose(actual_grad[0], expected_grad[0])
         assert np.isclose(actual_grad[1], expected_grad[1])
-
-
-class TestPreprocessingIntegration:
-    """Test preprocess produces output that can be executed by the device."""
-
-    @pytest.mark.parametrize("max_workers", [None, 1, 2])
-    def test_preprocess_single_circuit(self, max_workers):
-        """Test integration between preprocessing and execution with numpy parameters."""
-
-        # pylint: disable=too-few-public-methods
-        class MyTemplate(qml.operation.Operation):
-            """Temp operator."""
-
-            num_wires = 2
-
-            # pylint: disable=missing-function-docstring
-            def decomposition(self):
-                return [
-                    qml.RX(self.data[0], self.wires[0]),
-                    qml.RY(self.data[1], self.wires[1]),
-                    qml.CNOT(self.wires),
-                ]
-
-        x = 0.928
-        y = -0.792
-        qscript = qml.tape.QuantumScript(
-            [MyTemplate(x, y, ("a", "b"))],
-            [qml.expval(qml.PauliY("a")), qml.expval(qml.PauliZ("a")), qml.expval(qml.PauliX("b"))],
-        )
-
-        dev = DefaultQubit(max_workers=max_workers)
-        tapes = tuple([qscript])
-        program, config = dev.preprocess()
-        tapes, pre_processing_fn = program(tapes)
-
-        assert len(tapes) == 1
-        execute_circuit = tapes[0]
-        assert qml.equal(execute_circuit[0], qml.RX(x, "a"))
-        assert qml.equal(execute_circuit[1], qml.RY(y, "b"))
-        assert qml.equal(execute_circuit[2], qml.CNOT(("a", "b")))
-        assert qml.equal(execute_circuit[3], qml.expval(qml.PauliY("a")))
-        assert qml.equal(execute_circuit[4], qml.expval(qml.PauliZ("a")))
-        assert qml.equal(execute_circuit[5], qml.expval(qml.PauliX("b")))
-
-        results = dev.execute(tapes, config)
-        assert len(results) == 1
-        assert len(results[0]) == 3
-
-        processed_results = pre_processing_fn(results)
-        processed_result = processed_results[0]
-        assert len(processed_result) == 3
-        assert qml.math.allclose(processed_result[0], -np.sin(x) * np.sin(y))
-        assert qml.math.allclose(processed_result[1], np.cos(x))
-        assert qml.math.allclose(processed_result[2], np.sin(y))
-
-    @pytest.mark.parametrize("max_workers", [None, 1, 2])
-    def test_preprocess_batch_circuit(self, max_workers):
-        """Test preprocess integrates with default qubit when we start with a batch of circuits."""
-
-        # pylint: disable=too-few-public-methods
-        class CustomIsingXX(qml.operation.Operation):
-            """Temp operator."""
-
-            num_wires = 2
-
-            # pylint: disable=missing-function-docstring
-            def decomposition(self):
-                return [qml.IsingXX(self.data[0], self.wires)]
-
-        x = 0.692
-
-        measurements1 = [qml.density_matrix("a"), qml.vn_entropy("a")]
-        qs1 = qml.tape.QuantumScript([CustomIsingXX(x, ("a", "b"))], measurements1)
-
-        y = -0.923
-
-        with qml.queuing.AnnotatedQueue() as q:
-            qml.PauliX(wires=1)
-            m_0 = qml.measure(1)
-            qml.cond(m_0, qml.RY)(y, wires=0)
-            qml.expval(qml.PauliZ(0))
-
-        qs2 = qml.tape.QuantumScript.from_queue(q)
-
-        initial_batch = [qs1, qs2]
-
-        dev = DefaultQubit(max_workers=max_workers)
-
-        program, config = dev.preprocess()
-        batch, pre_processing_fn = program(initial_batch)
-
-        results = dev.execute(batch, config)
-        processed_results = pre_processing_fn(results)
-
-        assert len(processed_results) == 2
-        assert len(processed_results[0]) == 2
-
-        expected_density_mat = np.array([[np.cos(x / 2) ** 2, 0], [0, np.sin(x / 2) ** 2]])
-        assert qml.math.allclose(processed_results[0][0], expected_density_mat)
-
-        eig_1 = (1 + np.sqrt(1 - 4 * np.cos(x / 2) ** 2 * np.sin(x / 2) ** 2)) / 2
-        eig_2 = (1 - np.sqrt(1 - 4 * np.cos(x / 2) ** 2 * np.sin(x / 2) ** 2)) / 2
-        eigs = [eig_1, eig_2]
-        eigs = [eig for eig in eigs if eig > 0]
-
-        expected_entropy = -np.sum(eigs * np.log(eigs))
-        assert qml.math.allclose(processed_results[0][1], expected_entropy)
-
-        expected_expval = np.cos(y)
-        assert qml.math.allclose(expected_expval, processed_results[1])
-
-    def test_preprocess_defer_measurements(self, mocker):
-        """Test that a QNode with mid-circuit measurements is transformed
-        using defer_measurements."""
-        dev = DefaultQubit()
-
-        tape = qml.tape.QuantumScript(
-            [qml.PauliX(0), qml.measurements.MidMeasureMP(qml.wires.Wires(0))],
-            [qml.expval(qml.PauliZ(0))],
-        )
-        spy = mocker.spy(qml, "defer_measurements")
-
-        program, _ = dev.preprocess()
-        program([tape])
-        spy.assert_called_once()
 
 
 class TestRandomSeed:
@@ -1852,24 +1620,21 @@ class TestClassicalShadows:
             assert np.all(np.logical_or(np.logical_or(r[1] == 0, r[1] == 1), r[1] == 2))
 
 
-class TestDynamicType:
-    """Tests the compatibility with dynamic type classes such as `qml.Projector`."""
+@pytest.mark.parametrize("n_wires", [1, 2, 3])
+@pytest.mark.parametrize("max_workers", [None, 1, 2])
+def test_projector_dynamic_type(max_workers, n_wires):
+    """Test that qml.Projector yields the expected results for both of its subclasses."""
+    wires = list(range(n_wires))
+    dev = DefaultQubit(max_workers=max_workers)
+    ops = [qml.adjoint(qml.Hadamard(q)) for q in wires]
+    basis_state = np.zeros((n_wires,))
+    state_vector = np.zeros((2**n_wires,))
+    state_vector[0] = 1
 
-    @pytest.mark.parametrize("n_wires", [1, 2, 3])
-    @pytest.mark.parametrize("max_workers", [None, 1, 2])
-    def test_projector(self, max_workers, n_wires):
-        """Test that qml.Projector yields the expected results for both of its subclasses."""
-        wires = list(range(n_wires))
-        dev = DefaultQubit(max_workers=max_workers)
-        ops = [qml.adjoint(qml.Hadamard(q)) for q in wires]
-        basis_state = np.zeros((n_wires,))
-        state_vector = np.zeros((2**n_wires,))
-        state_vector[0] = 1
-
-        for state in [basis_state, state_vector]:
-            qs = qml.tape.QuantumScript(ops, [qml.expval(qml.Projector(state, wires))])
-            res = dev.execute(qs)
-            assert np.isclose(res, 1 / 2**n_wires)
+    for state in [basis_state, state_vector]:
+        qs = qml.tape.QuantumScript(ops, [qml.expval(qml.Projector(state, wires))])
+        res = dev.execute(qs)
+        assert np.isclose(res, 1 / 2**n_wires)
 
 
 class TestIntegration:
