+++ tests/devices/qubit/test_adjoint_jacobian.py
@@ -17,7 +17,11 @@ import pennylane as qml
 from pennylane.devices.qubit import adjoint_jacobian, adjoint_jvp, adjoint_vjp
 from pennylane.tape import QuantumScript
 import pennylane.numpy as np
-from pennylane.devices.qubit.preprocess import validate_and_expand_adjoint
+
+
+def adjoint_ops(op: qml.operation.Operator) -> bool:
+    """Specify whether or not an Operator is supported by adjoint differentiation."""
+    return op.num_params == 0 or op.num_params == 1 and op.has_generator
 
 
 class TestAdjointJacobian:
@@ -30,10 +34,7 @@ class TestAdjointJacobian:
         )
         qs.trainable_params = {0, 1}
 
-        qs_valid, _ = validate_and_expand_adjoint(qs)
-        qs_valid = qs_valid[0]
-
-        calculated_val = adjoint_jacobian(qs_valid)
+        calculated_val = adjoint_jacobian(qs)
 
         tapes, fn = qml.gradients.finite_diff(qs)
         results = tuple(qml.devices.qubit.simulate(t) for t in tapes)
@@ -48,17 +49,12 @@ class TestAdjointJacobian:
 
         prep_op = qml.StatePrep(np.array([1.0, -1.0], requires_grad=False) / np.sqrt(2), wires=0)
         qs = QuantumScript(
-            ops=[G(theta, wires=[0])], measurements=[qml.expval(qml.PauliZ(0))], prep=[prep_op]
+            ops=[prep_op, G(theta, wires=[0])], measurements=[qml.expval(qml.PauliZ(0))]
         )
 
         qs.trainable_params = {1}
-        qs_valid, _ = validate_and_expand_adjoint(qs)
-        qs_valid = qs_valid[0]
 
-        qs_valid.trainable_params = {1}
-
-        calculated_val = adjoint_jacobian(qs_valid)
-
+        calculated_val = adjoint_jacobian(qs)
         # compare to finite differences
         tapes, fn = qml.gradients.finite_diff(qs)
         results = tuple(qml.devices.qubit.simulate(t) for t in tapes)
@@ -75,18 +71,17 @@ class TestAdjointJacobian:
         prep_op = qml.StatePrep(np.array([1.0, -1.0], requires_grad=False) / np.sqrt(2), wires=0)
 
         qs = QuantumScript(
-            ops=[qml.Rot(*params, wires=[0])],
+            ops=[prep_op, qml.Rot(*params, wires=[0])],
             measurements=[qml.expval(qml.PauliZ(0))],
-            prep=[prep_op],
         )
 
         qs.trainable_params = {1, 2, 3}
-        qs_valid, _ = validate_and_expand_adjoint(qs)
-        qs_valid = qs_valid[0]
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
+        qs = qs_valid[0]
 
-        qs_valid.trainable_params = {1, 2, 3}
+        qs.trainable_params = {1, 2, 3}
 
-        calculated_val = adjoint_jacobian(qs_valid)
+        calculated_val = adjoint_jacobian(qs)
 
         # compare to finite differences
         tapes, fn = qml.gradients.finite_diff(qs)
@@ -120,7 +115,7 @@ class TestAdjointJacobian:
         qs = QuantumScript(ops, measurements)
         qs.trainable_params = set(range(1, 1 + op.num_params))
 
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         qs_valid.trainable_params = set(range(1, 1 + op.num_params))
@@ -143,7 +138,7 @@ class TestAdjointJacobian:
             [qml.RX(params[0], wires=0), qml.RX(params[1], wires=1), qml.RX(params[2], wires=2)],
             [qml.expval(qml.PauliZ(idx)) for idx in range(3)],
         )
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         # circuit jacobians
@@ -177,7 +172,7 @@ class TestAdjointJacobian:
             [MyOp(p, w) for p, w in zip(params, [0, 1, 2])],
             [qml.expval(qml.PauliZ(idx)) for idx in range(3)],
         )
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         # circuit jacobians
@@ -198,7 +193,7 @@ class TestAdjointJacobian:
         )
 
         qs.trainable_params = {1, 2, 3}
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         qs_valid.trainable_params = {1, 2, 3}
@@ -223,13 +218,12 @@ class TestAdjointJacobian:
         x, y, z = [0.5, 0.3, -0.7]
 
         qs = QuantumScript(
-            [qml.RX(0.4, wires=[0]), qml.Rot(x, y, z, wires=[0]), qml.RY(-0.2, wires=[0])],
+            [prep_op, qml.RX(0.4, wires=[0]), qml.Rot(x, y, z, wires=[0]), qml.RY(-0.2, wires=[0])],
             [qml.expval(qml.PauliZ(0))],
-            [prep_op],
         )
 
         qs.trainable_params = {2, 3, 4}
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         qs_valid.trainable_params = {2, 3, 4}
@@ -259,7 +253,7 @@ class TestAdjointJacobian:
         )
 
         qs.trainable_params = {0, 1, 2}
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         qs_valid.trainable_params = {0, 1, 2}
@@ -290,7 +284,7 @@ class TestAdjointJacobian:
         )
 
         qs.trainable_params = {0, 1, 2}
-        qs_valid, _ = validate_and_expand_adjoint(qs)
+        qs_valid, _ = qml.devices.preprocess.decompose(qs, adjoint_ops)
         qs_valid = qs_valid[0]
 
         res = adjoint_jacobian(qs_valid)
