+++ tests/test_return_types_dq2.py
@@ -1225,7 +1225,9 @@ class TestDeviceNewUnits:
 
         tape = qml.tape.QuantumScript.from_queue(q)
         dev = qml.device("default.qubit", wires=3)
-        with pytest.raises(qml.DeviceError, match="Analytic circuits must only contain"):
+        with pytest.raises(
+            qml.DeviceError, match="not accepted for analytic simulation on default.qubit"
+        ):
             program, _ = dev.preprocess()
             qml.execute(tapes=[tape], device=dev, gradient_fn=None, transform_program=program)
